requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.0
greenlet==3.2.4
idna==3.10
kiwisolver==1.4.9
lxml==6.0.1
matplotlib==3.10.6
mysql-connector-python==9.4.0
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
PyMySQL==1.1.2
pyparsing==3.2.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
seaborn==0.13.2
setuptools==59.6.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
tweepy==4.16.0
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0